feat: tests rejigged into new structure. starting to add travis
