+++ .travis.yml
@@ -1,0 +1,27 @@
+language: python
+sudo: required
+dist: xenial
+python: "3.7"
+
+# set up the image with system requirements
+addons:
+  apt:
+    update: true
+    packages:
+      - zsh
+      - make
+      - unzip
+
+# do the install, but only almost all because otherwise it takes hours
+install:
+  - make
+
+# check all branches
+branches:
+  only:
+#    - gh-pages
+    - /.*/
+
+# skip tests for now
+script:
+  - pytest
