+++ tox.ini
@@ -1,0 +1,4 @@
+[pytest]
+rootdir=./
+filterwarnings =
+    ignore::DeprecationWarning
